+++ Packages/com.unity.inputsystem/InputSystem/Plugins/PlayerInput/PlayerInputManager.cs
@@ -166,6 +166,20 @@ namespace UnityEngine.InputSystem
             }
         }
 
+        /// <summary>
+        /// The input action that a player must trigger to join the game.
+        /// </summary>
+        /// <remarks>
+        /// If the join action is a reference to an existing input action, it will be cloned when the PlayerInputManager
+        /// is enabled. This avoids the situation where the join action can become disabled after the first user joins which
+        /// can happen when the join action is the same as a player in-game action. When a player joins, input bindings from
+        /// devices other than the device they joined with are disabled. If the join action had a binding for keyboard and one
+        /// for gamepad for example, and the first player joined using the keyboard, the expectation is that the next player
+        /// could still join by pressing the gamepad join button. Without the cloning behavior, the gamepad input would have
+        /// been disabled.
+        ///
+        /// For more details about joining behavior, see <see cref="PlayerInput"/>.
+        /// </remarks>
         public InputActionProperty joinAction
         {
             get => m_JoinAction;
@@ -528,6 +542,15 @@ namespace UnityEngine.InputSystem
             {
                 Debug.LogWarning("Multiple PlayerInputManagers in the game. There should only be one PlayerInputManager", this);
                 return;
+            }
+
+            // if the join action is a reference, clone it so we don't run into problems with the action being disabled by
+            // PlayerInput when devices are assigned to individual players
+            if (joinAction.reference != null && joinAction.action?.actionMap?.asset != null)
+            {
+                var inputActionAsset = Instantiate(joinAction.action.actionMap.asset);
+                var inputActionReference = InputActionReference.Create(inputActionAsset.FindAction(joinAction.action.name));
+                joinAction = new InputActionProperty(inputActionReference);
             }
 
             // Join all players already in the game.